requirements.txt
bleak>=0.21.0
Pillow>=10.0.0
click>=8.1.0
